+++ pkg/util/k8sutil/util.go
@@ -52,14 +52,10 @@ const (
 	LabelKeyArangoActive = "deployment.arangodb.com/active"
 	// LabelValueArangoActive is the value of the label used to mark members as active.
 	LabelValueArangoActive = "true"
-	// LabelKeyArangoMLStatefulSet is the key of the label used to define k8s StatefulSet for ML Extension
-	LabelKeyArangoMLStatefulSet = "ml.arangodb.com/statefulset"
 	// LabelKeyAnalyticsGAEStatefulSet is the key of the label used to define k8s StatefulSet for Analytics GAE
 	LabelKeyAnalyticsGAEStatefulSet = "analytics.arangodb.com/statefulset"
 	// AppName is the value for the "app" label
 	AppName = "arangodb"
-	// AppArangoML is the value for the "app" label
-	AppArangoML = "arangoml"
 	// AppAnalytics is the value for the "app" label
 	AppAnalytics = "analytics"
 )